fix: 회원가입 API 경로 변경
