+++ src/main/java/shop/wannab/frontservice/auth/service/AuthClient.java
@@ -8,6 +8,8 @@ import shop.wannab.frontservice.auth.controller.request.LoginRequest;
 import shop.wannab.frontservice.auth.controller.request.ReissueRequest;
 import shop.wannab.frontservice.auth.controller.response.LoginResponse;
 import shop.wannab.frontservice.auth.controller.response.ReissueResponse;
+import shop.wannab.frontservice.user.dto.UserCreateRequest;
+import shop.wannab.frontservice.user.dto.UserPageResponse;
 
 @FeignClient(name = "gateway", path = "/user-service", contextId = "authClient")
 public interface AuthClient {
@@ -18,4 +20,6 @@ public interface AuthClient {
     @PostMapping("/api/auth/reissue")
     ResponseEntity<ReissueResponse> reissue(@RequestBody ReissueRequest request);
 
+    @PostMapping("/api/auth/signup")
+    ResponseEntity<UserPageResponse> createUser(@RequestBody UserCreateRequest dto);
 }
