+++ src/main/java/shop/wannab/frontservice/auth/service/AuthService.java
@@ -3,9 +3,15 @@ package shop.wannab.frontservice.auth.service;
 import io.jsonwebtoken.JwtException;
 import shop.wannab.frontservice.auth.controller.request.LoginRequest;
 import shop.wannab.frontservice.auth.controller.response.LoginResponse;
+import shop.wannab.frontservice.user.dto.UserCreateForm;
 
 public interface AuthService {
-    public LoginResponse login(LoginRequest request);
+    LoginResponse login(LoginRequest request);
 
-    public String validAccessToken(String accessToken, String refreshToken) throws JwtException;
+    String validAccessToken(String accessToken, String refreshToken) throws JwtException;
+
+    /**
+     * 회원 등록 Create
+     */
+    String createUser(UserCreateForm dto);
 }
