+++ src/main/java/shop/wannab/frontservice/auth/service/Impl/AuthServiceImpl.java
@@ -2,6 +2,7 @@ package shop.wannab.frontservice.auth.service.Impl;
 
 import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import shop.wannab.frontservice.auth.controller.request.LoginRequest;
@@ -10,6 +11,9 @@ import shop.wannab.frontservice.auth.controller.response.LoginResponse;
 import shop.wannab.frontservice.auth.controller.response.ReissueResponse;
 import shop.wannab.frontservice.auth.service.AuthClient;
 import shop.wannab.frontservice.auth.service.AuthService;
+import shop.wannab.frontservice.user.dto.UserCreateForm;
+import shop.wannab.frontservice.user.dto.UserCreateRequest;
+import shop.wannab.frontservice.user.dto.UserPageResponse;
 import shop.wannab.frontservice.utils.JwtUtils;
 
 @Service
@@ -43,6 +47,26 @@ public class AuthServiceImpl implements AuthService {
             return reissue.getBody().accessToken();
 
         throw new JwtException("AccessToken 재발급 중 예외 발생");
+    }
+
+    @Override
+    public String createUser(UserCreateForm userCreateForm) {
+        UserCreateRequest request = new UserCreateRequest(
+                userCreateForm.userId(),
+                userCreateForm.password(),
+                userCreateForm.username(),
+                userCreateForm.email(),
+                userCreateForm.phone(),
+                userCreateForm.birthday()
+        );
+
+        ResponseEntity<UserPageResponse> response = authClient.createUser(request);
+        switch (response.getStatusCode()) {
+            case HttpStatus.CREATED -> { return "success"; }
+            case HttpStatus.FORBIDDEN -> { return "허가되지 않은 요청입니다."; }
+            case HttpStatus.BAD_REQUEST -> { return "중복된 ID 입니다."; }
+            default -> throw new RuntimeException("예상치 못한 응답입니다: " + response.getStatusCode());
+        }
     }
 
 }
