+++ src/main/java/shop/wannab/frontservice/user/service/UserService.java
@@ -1,17 +1,9 @@
 package shop.wannab.frontservice.user.service;
 
-import shop.wannab.frontservice.user.dto.UserCreateForm;
 import shop.wannab.frontservice.user.dto.UserPageResponse;
-
-
 import shop.wannab.frontservice.user.dto.UserUpdateRequest;
 
 public interface UserService {
-
-    /**
-     * 회원 등록 Create
-     */
-    String createUser(UserCreateForm dto);
 
     /**
      * 회원 단건 조회 Read
